requirements.txt
RPi.GPIO
pygame
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
